OpenAPI: Update build and config files
This uses the newest version of openapi instead of swagger to build
the api. Swagger was on php v5, but this is on php 7.4.

In the newest unreleased version of openapi, they fix a deprecation
with php 8.4. That is why we are switching to openapi instead of
swagger which doesn't support php 8.4.